requirements.txt
fastapi==0.100.1
uvicorn==0.26.0
asyncpg==0.29.0
python-dotenv==1.0.0
requests==2.32.3
python-jose==3.4.0
Jinja2==3.1.2
pydantic==2.5.3
python-multipart
passlib[bcrypt]